Version: 0.151
Performed a LOT of optimizations and improved performance substantially
Reduced precision of textures used in fbo's to 16f
Improved SSAO performance and several other parts of the pipeline.

+++ src/Systems/Graphics/FX Techniques/Bloom_Tech.cpp
@@ -39,7 +39,7 @@ Bloom_Tech::Bloom_Tech(EnginePackage * enginePackage, Lighting_FBO * lightingFBO
 
 	glCreateFramebuffers(1, &m_fbo);
 	glCreateTextures(GL_TEXTURE_2D, 1, &m_texture);
-	glTextureImage2DEXT(m_texture, GL_TEXTURE_2D, 0, GL_RGB32F, m_renderSize.x, m_renderSize.y, 0, GL_RGB, GL_FLOAT, NULL);
+	glTextureImage2DEXT(m_texture, GL_TEXTURE_2D, 0, GL_RGB16F, m_renderSize.x, m_renderSize.y, 0, GL_RGB, GL_FLOAT, NULL);
 	glTextureParameteri(m_texture, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 	glTextureParameteri(m_texture, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 	glTextureParameteri(m_texture, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
@@ -49,7 +49,7 @@ Bloom_Tech::Bloom_Tech(EnginePackage * enginePackage, Lighting_FBO * lightingFBO
 
 	glCreateTextures(GL_TEXTURE_2D, 2, m_texturesGB);
 	for (int x = 0; x < 2; ++x) {
-		glTextureImage2DEXT(m_texturesGB[x], GL_TEXTURE_2D, 0, GL_RGB32F, m_renderSize.x, m_renderSize.y, 0, GL_RGB, GL_FLOAT, NULL);
+		glTextureImage2DEXT(m_texturesGB[x], GL_TEXTURE_2D, 0, GL_RGB16F, m_renderSize.x, m_renderSize.y, 0, GL_RGB, GL_FLOAT, NULL);
 		glTextureParameteri(m_texturesGB[x], GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 		glTextureParameteri(m_texturesGB[x], GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 		glTextureParameteri(m_texturesGB[x], GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
@@ -93,8 +93,8 @@ void Bloom_Tech::setBloomStrength(const int & strength)
 void Bloom_Tech::resize(const vec2 & size)
 {
 	m_renderSize = size;
-	glTextureImage2DEXT(m_texture, GL_TEXTURE_2D, 0, GL_RGB32F, m_renderSize.x, m_renderSize.y, 0, GL_RGB, GL_FLOAT, NULL);
+	glTextureImage2DEXT(m_texture, GL_TEXTURE_2D, 0, GL_RGB16F, m_renderSize.x, m_renderSize.y, 0, GL_RGB, GL_FLOAT, NULL);
 	glNamedFramebufferTexture(m_fbo, GL_COLOR_ATTACHMENT0, m_texture, 0);
 	for (int x = 0; x < 2; ++x) 
-		glTextureImage2DEXT(m_texturesGB[x], GL_TEXTURE_2D, 0, GL_RGB32F, m_renderSize.x, m_renderSize.y, 0, GL_RGB, GL_FLOAT, NULL);
+		glTextureImage2DEXT(m_texturesGB[x], GL_TEXTURE_2D, 0, GL_RGB16F, m_renderSize.x, m_renderSize.y, 0, GL_RGB, GL_FLOAT, NULL);
 }
